Moved partial class to own file
